requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pytest==7.4.3
requests==2.31.0
psutil==5.9.6
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
cryptography==41.0.7
PyJWT==2.8.0
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
numpy==1.25.2
pandas==2.1.3
scikit-learn==1.3.2
docker==6.1.3
kubernetes==28.1.0
prometheus-client==0.19.0
bandit==1.7.5
safety==2.3.5
black==23.11.0
flake8==6.1.0
